App Deployment/requirements.txt
flask>=3.0.0
pandas>=2.1.0
scikit-learn>=1.3.0
sastrawi>=1.0.1
spacy>=3.7.0
numpy>=1.24.0
gradio>=4.25.0
python-dotenv>=1.0.0
requests>=2.31.0
gensim>=4.3.0
gunicorn>=21.2.0
setuptools>=68.0.0
Werkzeug>=3.0.0
